feat: 검색 결과 표시 추가
