+++ src/app/search/search-empty.tsx
@@ -1,0 +1,7 @@
+export function SearchEmpty() {
+  return (
+    <div className="flex h-60 items-center justify-center text-gray-500">
+      검색어를 입력해주세요
+    </div>
+  );
+}
