+++ src/app/search/search-loading.tsx
@@ -1,0 +1,7 @@
+export function SearchLoading() {
+  return (
+    <div className="flex h-60 items-center justify-center">
+      <div className="text-gray-500">검색 중...</div>
+    </div>
+  );
+}
